requirements.txt
asgiref==3.1.1
asn1crypto==0.24.0
async-timeout==3.0.1
attrs==19.1.0
autobahn==19.3.3
Automat==0.7.0
certifi==2019.3.9
cffi==1.11.5
channels==2.2.0
chardet==3.0.4
constantly==15.1.0
cryptography==2.5
daphne==2.3.0
Django==2.1.5
django-cors-headers==2.4.0
django-filter==2.1.0
django-rest-knox==3.6.0
django-uuidfield==0.5.0
djangorestframework==3.9.1
drf-nested-routers==0.91
entrypoints==0.3
fcm-django==0.2.21
flake8==3.7.7
googlemaps==3.0.2
hyperlink==19.0.0
idna==2.8
incremental==17.5.0
isort==4.3.4
mccabe==0.6.1
Pillow==5.4.1
psycopg2==2.7.7
pycodestyle==2.5.0
pycparser==2.19
pyfcm==1.4.5
pyflakes==2.1.1
PyHamcrest==1.9.0
pyOpenSSL==19.0.0
pytz==2018.9
requests==2.21.0
requests-toolbelt==0.9.1
six==1.12.0
Twisted==19.2.0
txaio==18.8.1
urllib3==1.24.1
zope.interface==4.6.0
